pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wattplan"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
  "numpy",
  "pydantic>=2",
  "highspy",
]

[project.optional-dependencies]
test = [
  "pytest-homeassistant-custom-component==0.13.317",
]

[tool.pytest.ini_options]
addopts = "-ra"
pythonpath = ["."]
testpaths = [
  "tests/integration",
  "tests/optimizer",
]
asyncio_mode = "auto"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["custom_components*"]
namespaces = true

[tool.setuptools.package-data]
"custom_components.wattplan" = ["*.json", "*.yaml", "*.png", "translations/*.json"]
